requirements.txt
# Core dependencies
pyyaml>=6.0
colorlog>=6.0.0
tomli>=2.0.0
tomli-w>=1.0.0

# Server components
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.0.0
aiofiles>=23.2.1
websockets>=10.0

# Remote configuration client
requests>=2.31.0
websocket-client>=1.8.0

# Schema validation
jsonschema>=4.0.0
rfc3987>=1.3.8